requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
moviepy==1.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.2
imageio==2.31.1
imageio-ffmpeg==0.4.8
requests==2.31.0